# docs/platform-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Adaptive Coach Platform",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "summary": "Login",
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "summary": "Logout",
        "operationId": "logout_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete Account",
        "operationId": "delete_account_v1_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeletion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/me/export": {
      "get": {
        "summary": "Export",
        "operationId": "export_v1_me_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/plan": {
      "get": {
        "summary": "Get Plan",
        "operationId": "get_plan_v1_plan_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "summary": "Put Plan",
        "operationId": "put_plan_v1_plan_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/profile": {
      "get": {
        "summary": "Get Profile",
        "operationId": "get_profile_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "summary": "Put Profile",
        "operationId": "put_profile_v1_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/plan/history": {
      "get": {
        "summary": "History",
        "operationId": "history_v1_plan_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/activities": {
      "get": {
        "summary": "Activities",
        "operationId": "activities_v1_activities_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/activities/import": {
      "post": {
        "summary": "Import Activities",
        "operationId": "import_activities_v1_activities_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/activities/{provider}/{provider_id}": {
      "delete": {
        "summary": "Remove Activity",
        "operationId": "remove_activity_v1_activities__provider___provider_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/review/workout": {
      "get": {
        "summary": "Workout Review",
        "operationId": "workout_review_v1_review_workout_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/review/adjustments/preview": {
      "post": {
        "summary": "Preview",
        "operationId": "preview_v1_review_adjustments_preview_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/review/adjustments/{proposal_id}/apply": {
      "post": {
        "summary": "Apply",
        "operationId": "apply_v1_review_adjustments__proposal_id__apply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalAcceptance"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/integrations": {
      "get": {
        "summary": "Integrations",
        "operationId": "integrations_v1_integrations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/openapi.json": {
      "get": {
        "summary": "Schema",
        "operationId": "schema_v1_openapi_json_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AccountDeletion": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "password"
        ],
        "title": "AccountDeletion"
      },
      "ActivityImport": {
        "properties": {
          "ingestion_method": {
            "type": "string",
            "const": "client_import",
            "title": "Ingestion Method",
            "default": "client_import"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/ActivityRecord"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Activities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "activities"
        ],
        "title": "ActivityImport"
      },
...
      "Credentials": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Credentials"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Plan": {
        "properties": {
          "plan_name": {
            "type": "string",
            "minLength": 1,
            "title": "Plan Name"
          },
          "goal": {
            "type": "string",
            "title": "Goal",
            "default": ""
          },
          "athlete": {
            "additionalProperties": true,
            "type": "object",
            "title": "Athlete"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "workouts": {
            "items": {
              "$ref": "#/components/schemas/Workout"
            },
            "type": "array",
            "maxItems": 366,
            "minItems": 1,
            "title": "Workouts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_name",
          "workouts"
        ],
        "title": "Plan"
      },
      "PlanWrite": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "plan"
        ],
        "title": "PlanWrite"
      },
      "ProfileWrite": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "profile": {
            "$ref": "#/components/schemas/TrainingProfile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "profile"
        ],
        "title": "ProfileWrite"
      },
      "ProposalAcceptance": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "ProposalAcceptance"
      },
      "Registration": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          },
          "timezone": {
            "type": "string",
            "maxLength": 64,
            "title": "Timezone",
            "default": "Europe/Rome"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Registration"
      },
...
      "TrainingProfile": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "primary_sport": {
            "type": "string",
            "enum": [
              "running",
              "cycling",
              "both"
            ],
            "title": "Primary Sport"
          },
          "goal_type": {
            "type": "string",
            "enum": [
              "fitness",
              "consistency",
              "event",
              "personal_best"
            ],
            "title": "Goal Type"
          },
          "goal_description": {
            "type": "string",
            "maxLength": 1500,
            "minLength": 10,
            "title": "Goal Description"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "deadline_flexible": {
            "type": "boolean",
            "title": "Deadline Flexible",
            "default": true
          },
          "age_years": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 110.0,
                "minimum": 18.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Years"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 350.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 240.0,
                "minimum": 120.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "device_vendor": {
            "type": "string",
            "enum": [
              "none",
              "garmin",
              "coros",
              "suunto",
              "fitbit",
              "amazfit",
              "xiaomi",
              "apple",
              "other"
            ],
            "title": "Device Vendor"
          },
          "device_model": {
            "type": "string",
            "maxLength": 120,
            "title": "Device Model",
            "default": ""
          },
          "heart_rate_sensor": {
            "type": "boolean",
            "title": "Heart Rate Sensor",
            "default": false
          },
          "power_meter": {
            "type": "boolean",
            "title": "Power Meter",
            "default": false
          },
          "running_years": {
            "type": "number",
            "maximum": 80.0,
            "minimum": 0.0,
            "title": "Running Years"
          },
          "cycling_years": {
            "type": "number",
            "maximum": 80.0,
            "minimum": 0.0,
            "title": "Cycling Years"
          },
          "recent_running_km_week": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Recent Running Km Week"
          },
          "recent_cycling_km_week": {
            "type": "number",
            "maximum": 3000.0,
            "minimum": 0.0,
            "title": "Recent Cycling Km Week"
          },
          "experience_notes": {
            "type": "string",
            "maxLength": 2000,
            "title": "Experience Notes",
            "default": ""
          },
          "gym_sessions_week": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 0.0,
            "title": "Gym Sessions Week"
          },
          "gym_notes": {
            "type": "string",
            "maxLength": 1000,
            "title": "Gym Notes",
            "default": ""
          },
          "availability": {
            "items": {
              "$ref": "#/components/schemas/TrainingDay"
            },
            "type": "array",
            "maxItems": 7,
            "minItems": 1,
            "title": "Availability"
          },
          "best_performances": {
            "items": {
              "$ref": "#/components/schemas/BestPerformance"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Best Performances"
          },
          "constraints": {
            "type": "string",
            "maxLength": 1500,
            "title": "Constraints",
            "default": ""
          },
          "coaching_consent": {
            "type": "boolean",
            "title": "Coaching Consent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "primary_sport",
          "goal_type",
          "goal_description",
          "device_vendor",
          "running_years",
          "cycling_years",
          "recent_running_km_week",
          "recent_cycling_km_week",
          "gym_sessions_week",
          "availability",
          "coaching_consent"
        ],
        "title": "TrainingProfile"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Workout": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-zA-Z0-9_-]{1,80}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "sport": {
            "type": "string",
            "enum": [
              "running",
              "cycling",
              "rest",
              "manual"
            ],
            "title": "Sport"
          },
          "estimated_duration_min": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Duration Min"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "quality": {
            "type": "boolean",
            "title": "Quality",
            "default": false
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "name",
          "sport"
        ],
        "title": "Workout"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}